Version Final

+++ src/ficheros/Comandos.java
@@ -60,18 +60,24 @@ public class Comandos {
         help += "delete <nombre> -> Borra el fichero, si es un directorio borra todo su contenido y a si mismo\n";
         help += "close -> Cierra el programa\n";
         help += "clear -> Vacia la vista\n";
-        return help;
+        if (valores.length == 1) {
+            return help;
+        } else {
+            return "Demasiados argumentos";
+        }
     }
 
     public void cd() {
         if (valores.length == 1) {
-            System.out.println(System.getProperty("user.dir"));
+            System.out.println(ruta);
         } else if (valores.length == 2) {
             switch (valores[1]) {
                 case ".." -> ruta = ruta.getParent();
                 case "/" -> rutaAbsoluta();
                 default -> System.out.println(nombreDirectorio());
             }
+        } else if (valores.length >= 3) {
+            System.out.println("Demasiados argumentos");
         }
     }
 
@@ -99,16 +105,28 @@ public class Comandos {
     }
 
     public void mkdir() {
-        String nuevo = valores[1];
-        File d = new File(ruta.resolve(nuevo).toString());
-        if (!d.exists()) {
-            if (d.mkdir()) {
-                System.out.println("Directorio creado");
-            } else {
-                System.out.println("No se ha podido crear el directorio");
+        if (valores.length == 2) {
+            String nuevo = valores[1];
+            File d = new File(ruta.resolve(nuevo).toString());
+            if (!d.exists()) {
+                if (d.mkdir()) {
+                    System.out.println("Directorio creado");
+                } else {
+                    System.out.println("No se ha podido crear el directorio");
+                }
+            } else if (d.exists()) {
+                System.out.println("El directorio ya existe");
+                System.out.println("Quieres borrarlo?");
+                if (sc.nextLine().equalsIgnoreCase("si")) {
+                    deleteDirectorio(d);
+                    System.out.println("Fichero borrado");
+                }
+
             }
-        } else if (d.exists()) {
-            System.out.println("El directorio ya existe");
+        } else if (valores.length == 1) {
+            System.out.println("Falta el nombre del directorio");
+        } else if (valores.length >= 3) {
+            System.out.println("Demasiados argumentos");
         }
     }
 
@@ -130,10 +148,9 @@ public class Comandos {
                 System.out.println("Ruta absoluta: " + ruta.toAbsolutePath());
                 System.out.println("Raiz de la ruta: " + ruta.getRoot());
                 System.out.println("Numero de elementos de la ruta: " + ruta.getNameCount());
-                System.out.println("Espacio libre en la ruta: " + ruta.toFile().getFreeSpace());
-                System.out.println("Espacio total en la ruta: " + ruta.toFile().getTotalSpace());
-                System.out.println("Espacio utilizable en la ruta: " + ruta.toFile().getUsableSpace());
-                System.out.println("Ultima modificacion: " + ruta.toFile().lastModified());
+                System.out.println("Espacio libre en la ruta: " + ruta.toFile().getFreeSpace() + " bytes");
+                System.out.println("Espacio total en la ruta: " + ruta.toFile().getTotalSpace() + " bytes");
+                System.out.println("Espacio utilizable en la ruta: " + ruta.toFile().getUsableSpace() + " bytes");
                 System.out.println("Tamaño: " + ruta.toFile().length());
                 System.out.println("Es un directorio: " + ruta.toFile().isDirectory());
                 System.out.println("Es un fichero: " + ruta.toFile().isFile());
@@ -142,118 +159,166 @@ public class Comandos {
     }
 
     public void cat() {
-        String nombre = valores[1];
-        File f = new File(ruta.resolve(nombre).toString());
-        if (!f.exists()) {
-            System.out.println("Ese fichero no existe");
-        } else {
-            try {
-                FileReader fileLectura = new FileReader(f);
-                BufferedReader bufferLectura = new BufferedReader(fileLectura);
-                System.out.println("Leyendo fichero...");
-                int i;
-                while ((i = bufferLectura.read()) != -1) {
-                    System.out.print((char) i);
+        if (valores.length == 2) {
+            String nombre = valores[1];
+            File f = new File(ruta.resolve(nombre).toString());
+            if (!f.exists()) {
+                System.out.println("Ese fichero no existe");
+            } else {
+                try {
+                    FileReader fileLectura = new FileReader(f);
+                    BufferedReader bufferLectura = new BufferedReader(fileLectura);
+                    System.out.println("Leyendo fichero...");
+                    int i;
+                    while ((i = bufferLectura.read()) != -1) {
+                        System.out.print((char) i);
+                    }
+                    bufferLectura.close();
+                } catch (IOException e) {
+                    e.printStackTrace();
                 }
-                bufferLectura.close();
-            } catch (IOException e) {
-                e.printStackTrace();
             }
+        } else if (valores.length == 1) {
+            System.out.println("Falta el nombre del fichero");
+        } else if (valores.length >= 3) {
+            System.out.println("Demasiados argumentos");
         }
     }
 
     public void top() {
-        String nombre = valores[1];
-        System.out.println("Introduce el numero de lineas que quieres mostrar: ");
-        int lineas = sc.nextInt();
-        File f = new File(ruta.resolve(nombre).toString());
-        if (!f.exists()) {
-            System.out.println("Ese fichero no existe");
-        } else {
-            try {
-                FileReader fileLectura = new FileReader(f);
-                BufferedReader bufferLectura = new BufferedReader(fileLectura);
-                System.out.println("Leyendo fichero...");
-                int i;
-                int contador = 1;
-                while ((i = bufferLectura.read()) != -1) {
-                    if (contador <= lineas) {
-                        System.out.print((char) i);
-                        if ((char) i == "\n".charAt(0)) {
-                            contador++;
+        if (valores.length == 3) {
+            String nombre = valores[2];
+            int lineas = Integer.parseInt(valores[1]);
+            File f = new File(ruta.resolve(nombre).toString());
+            if (!f.exists()) {
+                System.out.println("Ese fichero no existe");
+            } else {
+                try {
+                    FileReader fileLectura = new FileReader(f);
+                    BufferedReader bufferLectura = new BufferedReader(fileLectura);
+                    System.out.println("Leyendo fichero...");
+                    int i;
+                    int contador = 1;
+                    while ((i = bufferLectura.read()) != -1) {
+                        if (contador <= lineas) {
+                            System.out.print((char) i);
+                            if ((char) i == "\n".charAt(0)) {
+                                contador++;
+                            }
                         }
                     }
+                    bufferLectura.close();
+                } catch (IOException e) {
+                    e.printStackTrace();
                 }
-                bufferLectura.close();
-            } catch (IOException e) {
-                e.printStackTrace();
             }
+        } else if (valores.length == 1) {
+            System.out.println("Falta el nombre del fichero");
+        } else if (valores.length == 2) {
+            System.out.println("Falta el numero de lineas");
+        } else if (valores.length >= 4) {
+            System.out.println("Demasiados argumentos");
         }
     }
 
     public void mkfile() {
-        String nombre = valores[1];
-        File f = new File(ruta.resolve(nombre).toString());
-        if (f.exists()) {
-            System.out.println("Ese fichero ya existe");
-            System.out.println("Quieres borrarlo?");
-            if (sc.nextLine().equals("Si")) {
-                f.delete();
-                System.out.println("Fichero borrado");
-            }
-        } else {
-            try {
-                if (f.createNewFile()) {
-                    System.out.println("Introduce el texto que contendra el fichero");
-                    String texto = sc.nextLine();
-                    FileWriter file = new FileWriter(f);
-                    BufferedWriter buffer = new BufferedWriter(file);
-                    buffer.write(texto);
-                    buffer.flush();
-                    buffer.close();
-                    System.out.println("Fichero creado");
-                } else {
-                    System.out.println("No se ha podido crear el archivo");
+        if (valores.length == 3) {
+            String nombre = valores[1];
+            File f = new File(ruta.resolve(nombre).toString());
+            if (f.exists()) {
+                System.out.println("Ese fichero ya existe");
+                System.out.println("Quieres borrarlo?");
+                if (sc.nextLine().equalsIgnoreCase("si")) {
+                    f.delete();
+                    System.out.println("Fichero borrado");
                 }
-            } catch (IOException e) {
-                e.printStackTrace();
+            } else {
+                try {
+                    if (f.createNewFile()) {
+                        String texto = valores[2];
+                        FileWriter file = new FileWriter(f);
+                        BufferedWriter buffer = new BufferedWriter(file);
+                        buffer.write(texto);
+                        buffer.flush();
+                        buffer.close();
+                        System.out.println("Fichero creado");
+                    } else {
+                        System.out.println("No se ha podido crear el archivo");
+                    }
+                } catch (IOException e) {
+                    e.printStackTrace();
+                }
             }
+        } else if (valores.length == 1) {
+            System.out.println("Falta el nombre del fichero");
+        } else if (valores.length == 2) {
+            System.out.println("Falta el texto del fichero");
+        } else if (valores.length >= 4) {
+            System.out.println("Demasiados argumentos");
         }
     }
 
     public void write() {
-        String nombre = valores[1];
-        String salida = "";
-        File f = new File(ruta.resolve(nombre).toString());
-        if (!f.exists()) {
-            System.out.println("Ese fichero no existe");
-        } else {
-            try {
-                System.out.println("Introduce el texto que contendra el fichero");
-                String texto = sc.nextLine();
-                FileReader fileLectura = new FileReader(f);
-                BufferedReader bufferLectura = new BufferedReader(fileLectura);
-                System.out.println("Leyendo fichero...");
-                int i = 0;
-                while ((i = bufferLectura.read()) != -1) {
-                    char c = (char) i;
-                    salida += c + "";
+        if (valores.length == 3) {
+            String nombre = valores[1];
+            String salida = "";
+            File f = new File(ruta.resolve(nombre).toString());
+            if (!f.exists()) {
+                System.out.println("Ese fichero no existe");
+            } else {
+                try {
+                    String texto = valores[2];
+                    FileReader fileLectura = new FileReader(f);
+                    BufferedReader bufferLectura = new BufferedReader(fileLectura);
+                    System.out.println("Leyendo fichero...");
+                    int i = 0;
+                    while ((i = bufferLectura.read()) != -1) {
+                        char c = (char) i;
+                        salida += c + "";
+                    }
+                    FileWriter file = new FileWriter(f);
+                    BufferedWriter buffer = new BufferedWriter(file);
+                    buffer.write(salida + " " + texto);
+                    buffer.flush();
+                    buffer.close();
+                    System.out.println("Fichero escrito");
+                } catch (IOException e) {
+                    e.printStackTrace();
                 }
-                FileWriter file = new FileWriter(f);
-                BufferedWriter buffer = new BufferedWriter(file);
-                buffer.write(salida + " " + texto);
-                buffer.flush();
-                buffer.close();
-                System.out.println("Fichero escrito");
-            } catch (IOException e) {
-                e.printStackTrace();
             }
+        } else if (valores.length == 1) {
+            System.out.println("Falta el nombre del fichero");
+        } else if (valores.length == 2) {
+            System.out.println("Falta el texto que quieres escribir");
+        } else if (valores.length >= 4) {
+            System.out.println("Demasiados argumentos");
         }
     }
 
     public void dir() {
-        String nombre = valores[1];
-        File f = new File(ruta.resolve(nombre).toString());
+        if (valores.length == 1) {
+            File f = new File(ruta.toString());
+            dirAccion(f);
+        } else if (valores.length == 2) {
+            switch (valores[1]) {
+                case "/" -> {
+                    Path rutaAbsoluta = Paths.get(valores[1]);
+                    File f = rutaAbsoluta.toFile();
+                    dirAccion(f);
+                }
+                default -> {
+                    String nombre = valores[1];
+                    File f = new File(ruta.resolve(nombre).toString());
+                    dirAccion(f);
+                }
+            }
+        }
+        else if (valores.length >= 3) {
+            System.out.println("Demasiados argumentos");
+        }
+    }
+
+    private void dirAccion(File f) {
         File[] lista = f.listFiles();
         if (lista != null) {
             System.out.println("Contenido de la ruta " + ruta + ":");
@@ -268,19 +333,29 @@ public class Comandos {
         }
     }
 
-    public void delete(){
-        String nombre = valores[1];
-        File f = new File(ruta.resolve(nombre).toString());
-        if (f.isDirectory()) {
-            deleteDirectorio(f);
-        } else {
-            f.delete();
+    public void delete() {
+        if (valores.length == 2) {
+            String nombre = valores[1];
+            File f = new File(ruta.resolve(nombre).toString());
+            if (f.exists()) {
+                if (f.isDirectory()) {
+                    deleteDirectorio(f);
+                } else {
+                    f.delete();
+                }
+            } else {
+                System.out.println("Ese fichero no existe");
+            }
+        } else if (valores.length == 1) {
+            System.out.println("Falta el nombre del fichero");
+        } else if (valores.length >= 3) {
+            System.out.println("Demasiados argumentos");
         }
     }
 
-    public static void deleteDirectorio(File direc) {
-        if (direc.exists()) {
-            File[] files = direc.listFiles();
+    public static void deleteDirectorio(File direct) {
+        if (direct.exists()) {
+            File[] files = direct.listFiles();
             if (files != null) {
                 for (File hijo : files) {
                     if (hijo.isDirectory()) {
@@ -290,18 +365,26 @@ public class Comandos {
                     }
                 }
             }
-            direc.delete();
+            direct.delete();
         }
     }
 
     public void close() {
-        System.out.println("Cerrando programa...");
-        System.exit(0);
+        if (valores.length == 1) {
+            System.out.println("Cerrando programa...");
+            System.exit(0);
+        } else if (valores.length >= 2) {
+            System.out.println("Demasiados argumentos");
+        }
     }
 
     public void clear() {
-        for (int i = 0; i < 50; i++) {
-            System.out.println("");
+        if (valores.length == 1) {
+            for (int i = 0; i < 50; i++) {
+                System.out.println("");
+            }
+        } else if (valores.length >= 2) {
+            System.out.println("Demasiados argumentos");
         }
     }
 }
